\name{newCountDataSetFromHTSeqCount}
\alias{newCountDataSetFromHTSeqCount}
\title{
Create a new CountDataSet from count files generated with htseq-count
}
\description{Use this function to start a DESeq analysis if you used htseq-count
to count your reads.}
\usage{
newCountDataSetFromHTSeqCount(sampleTable, directory = ".")
}
\arguments{
  \item{sampleTable}{
     A data frame with three or more columns. Each row describes one sample.
The first column is the sample name, the seond column the file name of
the count file generated by htseq-count, and the remaining columns are
sample meta data. If the meta data consists of only a single column
(i.e., three columns in total), this is used as 'condition' factor. 
}
  \item{directory}{
The directory relative to which the filenames are specified.
}
}
\value{
A CountDataSet object.}
\references{
See http://www-huber.embl.de/users/anders/HTSeq/ for htseq-count.}
\author{
Simon Anders}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{newCountDataSet}}}

